Support decompression of PDB files for pose_from_pdb and pose_from_file in pyrosetta.io namespace
